feat(react): optimize react file structure
