+++ packages/react/mobile-first.ts
@@ -1,13 +1,13 @@
 import Alert from '@opentiny/react-alert/src/mobile-first'
 import Button from '@opentiny/react-button/src/mobile-first'
-
-const components = [Alert, Button]
+import Icon from '@opentiny/react-icon/src/mobile-first'
 
 export const version = '1.0.0'
 
-export { Alert, Button }
+export { Alert, Button, Icon }
 
 export default {
   Alert,
-  Button
+  Button,
+  Icon
 } as any
