+++ packages/react/src/common/src/render-stack.ts
