+++ packages/react/src/common/src/resolve-props.ts
