+++ packages/react/src/icon/index.ts
