+++ packages/react/src/icon/package.json
