+++ packages/react/src/icon/src/check/index.ts
