+++ packages/react/src/icon/src/close/index.ts
