+++ packages/react/src/icon/src/error/index.ts
