+++ packages/react/src/icon/src/half-select/index.ts
