+++ packages/react/src/icon/src/loading/index.ts
